Add Auhtorization (Custom Policeis ,custom Authorizaiton Based on Resources)

+++ LosPollos.Application/Commands/Restaurants/DeleteCommands/DeleteRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using LosPollos.Application.Commands.Restaurants.CreateCommands;
+using LosPollos.Domain.Constant;
 using LosPollos.Domain.Entities;
 using LosPollos.Domain.Exceptions;
+using LosPollos.Domain.Interfaces;
 using LosPollos.Domain.Interfaces.Repository;
 using MediatR;
 using Microsoft.Extensions.Logging;
@@ -12,11 +14,13 @@ namespace LosPollos.Application.Commands.Restaurants.DeleteCommands
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly ILogger<DeleteRestaurantCommandHandler> _logger;
-       
-        public DeleteRestaurantCommandHandler(IUnitOfWork unitOfWork, ILogger<DeleteRestaurantCommandHandler> logger)
+        private readonly IRestaurantAuhtorizationServices _AuthServices;
+
+        public DeleteRestaurantCommandHandler(IUnitOfWork unitOfWork, ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantAuhtorizationServices authServices)
         {
             _unitOfWork = unitOfWork;
             _logger = logger;
+            _AuthServices = authServices;
         }
 
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -25,7 +29,8 @@ namespace LosPollos.Application.Commands.Restaurants.DeleteCommands
             var restaurant = await _unitOfWork.restaurantRepository.GetAsync(x => x.Id == request.Id);
             if (restaurant is null)
                 throw new NotFoundException(nameof(Resturant), request.Id.ToString());
-
+            if (!_AuthServices.Authorize(restaurant, ResourceOperation.Delete))
+                throw new ForbidException();
             await _unitOfWork.restaurantRepository.DeleteAsync(restaurant);
             await _unitOfWork.Save();
